requirements.txt
accelerate>=0.19.0
transformers>=4.30.0
diffusers>=0.18.0
torch>=2.0.0
Pillow>=9.0.0
tqdm>=4.65.0
bitsandbytes>=0.39.0
safetensors>=0.3.1
peft>=0.4.0
# For improved performance
xformers
# triton is optional and may not be available for all platforms 